+++ hermod-tracing-core/src/Hermod/Tracing/Tracer/EKG.hs
@@ -73,10 +73,10 @@ ekgTracer TraceConfig{tcMetricsPrefix} store = liftIO $ do
             let fullName = metricsPrefix <> name <> "_real"
             label <- modifyMVar rgsLabels (setFunc Metrics.createLabel fullName)
             Label.set label (showTReal theDouble)
-          PrometheusM name keyLabels -> do
+          LabelSetM name keyLabels -> do
             let fullName = metricsPrefix <> name
             label <- modifyMVar rgsLabels (setFunc Metrics.createLabel fullName)
-            Label.set label (presentPrometheusM keyLabels)
+            Label.set label (presentLabelSetM keyLabels)
           CounterM name mbInt -> do
             let fullName = metricsPrefix <> name <> "_counter"
             counter <- modifyMVar rgsCounters (setFunc Metrics.createCounter fullName)
@@ -97,8 +97,8 @@ ekgTracer TraceConfig{tcMetricsPrefix} store = liftIO $ do
             let rgsMap' = Map.insert name metric rgsMap
             pure (rgsMap', metric)
 
-    presentPrometheusM :: [(Text, Text)] -> Text
-    presentPrometheusM =
+    presentLabelSetM :: [(Text, Text)] -> Text
+    presentLabelSetM =
       label . map pair
       where
         label pairs = "{" <> intercalate "," pairs <> "} 1"